requirements.txt
kaggle
loguru
numpy==1.26.4
open_clip_torch==3.0.0
opencv-python==4.10.0.84
pandas==2.3.1
pillow==11.1.0
rich
scikit-learn==1.15.0
seaborn==0.13.2
tensorboard
torch
torchvision
transformers==4.44.0
tqdm
